requirements.txt
asgiref==3.7.2
autopep8==2.0.2
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
Django==5.0.1
djangorestframework==3.15.1
filelock==3.13.1
fonttools==4.45.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.2.0
pycodestyle==2.10.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.25.0
asgiref==3.3.4
Django==3.2.1
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
